KNL-1337 Refactor out ProviderMap into file.
ProviderMap doesn’t need to be an inner class and this reduces the size of a very large class.
